Structure to the Model with some package defining.
Trips contains Features (Calendar, Photoalbum, Phonebook, etc.) and Schedulable objects (Accommodation, Itinerary, Reservation, and more.)
Users contain Trips and a username/password, will need an encryption/decryption thing when storing file information.
Schedulable objects need to be able to export all their data as a HashMap whose Key set is the field of the data (Title, Description, etc.; generic terms for information) and the Values being the corresponding data. This is so that when drawing, there are consistent naming conventions and rules for displayed data (and for editing data when information is passed back in.)
Needs more work and fleshing out, but it's a start for day one.